refactor: move to switch between jsc and hermes
